pyproject.toml
[project]
name = "gen-ai"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "autogen>=0.12.1",
    "beautifulsoup4>=4.14.3",
    "crewai[litellm]>=1.14.3",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.2",
    "langchain-ollama>=1.1.0",
    "langgraph>=1.1.9",
    "litellm>=1.83.0",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "torchvision>=0.26.0",
]
